Release 0.28.1
This release is a bit different from the other ones since we
had to get around the following (impactful) commit:

https://github.com/apache/incubator-superset/commit/160e47720cdfbe69e0cc8c0ecffc63b7c5ad0b72

Python 3.6+ is not available on Debian Stretch (only on Buster,
that is the current testing) so we have to temporarily build
our own patched version of the superset wheel to get around this
issue. For this release it seems that no specific 3.6 construct
(like f-strings) has been used, so a revert of the above commit
should be enough to allow the code to run on Python 3.5.

Bug: T211605
Change-Id: Iafadac4a2ab8525b927b02313e398603b925c088