build_wheels.sh: use the system pip instead of the virtualenv's one
What:
Avoid installing any pip when creating the virtual environment,
and rely on the system's one.

Why:
As outlined in T236690#5697280, pip in a virtual environment is the
last upstream version. This may be a problem when building wheels
that contain c-code (like pyarrow), if Debian's pip is used
to deploy those wheels afterwards.

Bug: T236690
Change-Id: Ie5d5d738c9f88ad992c92ca77b17cc700b0b95fb