Update to Superset version 3.1.0
We update the build process so as to build from the upstream superset
source, rather than just installing packages from pypi.

Add the superset[trino,hive,postgresql,spark] extras in order to
generate a more complete set of dependencies.

In addition to this, we have temporarily applied the patch
to fix this issue: https://github.com/apache/superset/issues/25636

Bug: T335356
Change-Id: I9a47ae07a106a1481fdffe03f670ced4c46b262d