Improve the display of nested columns from presto
An issue has been identified with the way in which presto attempts to
expand nested columns in presto.

Three patches are applied to a pristine superset 3.1.0 checkout during
the build process.

The first fixes a regression in the display of schemas on presto.

The second fixes a bug in the display of nested schemas, which is behind
the PRESTO_EXPAND_DATA feature flag.

The third is a newly proposed feature, whereby nested data is expanded
to show column names as well as values in the cell display modal dialog.

We are coordinating with the upstream authors to get these patches
included in Superset. The first two already have PRs and the third is
under discussion. Links to these are included in the comments.

Bug: T340144
Change-Id: If824b1aa11b5f6db248a41faa370c5ae9e1d3d00